+++ shopify_sync/models/line_item.py
@@ -19,7 +19,7 @@ class LineItem(ShopifyResourceModel):
     name = models.CharField(max_length = 256)
     order = models.ForeignKey('shopify_sync.Order')
     price = models.DecimalField(max_digits = 10, decimal_places = 2)
-    product_id = models.IntegerField(null = True)
+    product_id = models.BigIntegerField(null = True)
     product_exists = models.BooleanField(default = True)
     properties = JSONField(default = empty_list, dump_kwargs = {'cls': ShopifyDjangoJSONEncoder})
     quantity = models.IntegerField()
@@ -30,7 +30,7 @@ class LineItem(ShopifyResourceModel):
     tax_lines = JSONField(default = empty_list, dump_kwargs = {'cls': ShopifyDjangoJSONEncoder})
     title = models.CharField(max_length = 256)
     total_discount = models.DecimalField(max_digits = 10, decimal_places = 2)
-    variant_id = models.IntegerField(null = True)
+    variant_id = models.BigIntegerField(null = True)
     variant_title = models.CharField(max_length = 256)
     vendor = models.CharField(max_length = 64)
